Severely optimized rendering, revise TODO list
The buffers for OpenGL-rendering (vertex buffers, etc.) were segmented
into individual objects to Severely optimize rendering, as a single set
of buffers isn't used for *all* rendering. This is means that video
memory is actually utilized.

The TODO list had a lot of points added to it, and more long term goals
were moved to a seperate section