Initiate transition to C
I have decided that this project should be written in C and not C++,
primarily due to performance concerns. And also because I quite simply
find C a much more pleasant language to write in.

This commit initiates the transition to C, and introduces (yet again)
some changes to the rendering pipeline. Also not that this commit is
complete, which is why I'm pushing it to a seperate branch.