Segfault fix, Object implementation, naming style
Or: Oh boy, what a mess

So I located the cause of the segfaulting-on-window-creation thing. Sort
of. Turns out, it only occurs when the program is compiled as a C
program, and not when it's compiled as a C++ one. I have no idea why,
but I guess I'll just have to deal with it. From this point on I will still
keep the code as C-compatible as possible, with some minor exceptions
regarding use of syntactic sugar.

I'm also reverting to the coordinate and size model for now, because I'd
prefer to focus on building the general structure of the program for
now, and add more advanced object logic later.

Lastly, the naming convention has changed slightly. Types now have
uppercase names.